+++ src/packages/media/media/recycle-bin/tree/index.ts
@@ -4,6 +4,6 @@ export {
 	UMB_MEDIA_RECYCLE_BIN_TREE_STORE_ALIAS,
 	UMB_MEDIA_RECYCLE_BIN_TREE_ALIAS,
 } from './constants.js';
-export { UMB_MEDIA_RECYCLE_BIN_TREE_STORE_CONTEXT } from './media-recycle-bin-tree.store.js';
+export { UMB_MEDIA_RECYCLE_BIN_TREE_STORE_CONTEXT } from './media-recycle-bin-tree.store.context-token.js';
 export { type UmbMediaRecycleBinTreeStore } from './media-recycle-bin-tree.store.js';
 export * from './types.js';
