+++ src/packages/media/media/recycle-bin/tree/media-recycle-bin-tree.store.ts
@@ -1,4 +1,4 @@
-import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
+import { UMB_MEDIA_RECYCLE_BIN_TREE_STORE_CONTEXT } from './media-recycle-bin-tree.store.context-token.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbUniqueTreeStore } from '@umbraco-cms/backoffice/tree';
 
@@ -20,7 +20,3 @@ export class UmbMediaRecycleBinTreeStore extends UmbUniqueTreeStore {
 }
 
 export { UmbMediaRecycleBinTreeStore as api };
-
-export const UMB_MEDIA_RECYCLE_BIN_TREE_STORE_CONTEXT = new UmbContextToken<UmbMediaRecycleBinTreeStore>(
-	'UmbMediaRecycleBinTreeStore',
-);
